Begin splitting up proptest code.
